Fix NPE on Root level types
If you provide a schema that does not have
either the Query or the Mutation root type
the script errors out.

This fix makes it so that it does not error
out. Instead it skips the missing type, or
if both types are missing, it will output
an empty entitlements map.